+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Unreleased
 * Full support for constraint auxiliary variables (lin eq, nonlin eq, line ineq) (done).
 * Elimination of obscure variables "voltage magnitude deviation" (done).
 * Elimination of obscure variabes "voltage magnitude violation", "tap ratio deviation", "susceptance deviation" (done).
+* Improved setup.py without argparse that relies of existing build_ext commands for custom builds (done).
+* Upper and lower bounds for constraint extra variables (done).
 * Load constant power factor.
 
 Version 1.2.9